move discover mappers with all the other mappers
